Handle relocation of notifier for copying attributes. (#99)
Notifier 'NOTIFY_ADMIN_PRODUCT_COPY_TO_ATTRIBUTES' was requested to be
moved outside of the specific grouping for checking if a product has
attributes.  This was agreed upon so had to ensure that the code would
still work properly. Wanted to be sure to provide/force consistency
where able.  Unfortunately the notifier is in code that doesn't require
modification for operation so it may/may not be inside/outside the
code grouping depending on which version is loaded, etc...

Fixes #98