Prevent trying non-existent file
A default install of Zen Cart 1.5.7x will include a classes file in the admin with content indicating that the file is no longer needed; however, for those that are familiar with that may remove the file.  If the file is removed, the original suggested code operation would stop in attempting to process the first array element. This causes the loop to continue if the attempted class file doesn't exist.