Potentially remove edit of most admin files
The goal of this is to reduce/remove the need for edits to the associated admin files. This captures the possibility of displaying the customid in each of the admin files of `invoice`, `packingslip`, and `orders` by prepending the customid to the option name if each option name is to have a customid or by appending to the model if there is only one customid set of text to be displayed (single attribute or where a single variant is comprised of multiple attributes).  Only one edit to a Zen Cart 1.5.7c fileset remains which is a notifier within admin/packingslip.php.

There is some excess code included which was introduced for the invoice. When looking to expand functionality and realizing the limitations of the existing invoice output, the excess code was prevented from executing by commenting out the two associated notifiers and was then transferred into the existing code that populates the admin's order class based on default Zen Cart code.

The additional admin/packingslip.php notifier had been added for additional image display that was at one point previously adapted for other use.

Additionally, in review of the code for customid, an anomaly was discovered that could prevent the display of the customid, therefore the sequence was modified to instead of jointly step through an "imaginary" order list of product and attempt to compare against the actual list in the order to just step through the order itself in a "generic" way (not requiring a sequential list of product, but instead any/all product) and obtain data based off of that information.

Also refactored the function to provide early continue on loops and improve readability of the code within this section.