Expand Account History determination of customid and refactor
Change the use of for to foreach to reduce potential for issues if information
fed to the code is not sequentially placed.

Added comments to the code, expanded the data preparation for account history
instead of just the one single "combination" customid, the ability to place
the customid with the associated attribute has also been incorporated.