Refactor shopping_cart information determination
Switch to using foreach instead of for in case product are removed from
the cart without refactoring the numbering.
Support early exit from/through looping to simplify reading of and understanding
of the code.

Mostly this shifted the code to the left, because parts of the if/if else/else
structure were removed and placed at the top of the loop.