Refactor setCheckStockParams to remove large sweeping "if" statement.
While this may come back to haunt as in there may still be a need for
some "large" if statement, at this time it is not necessary. Further, there
are/were additional actions that could be taken in absence of existing data
to support the follow through of this code section. Basically, since it
may be necessary to review information about an order, that the order needs
to be "creatable" on demand.  Further, in such an on demand situation there
may be variables sought that don't exist, so added some strict control logic
in to prevent generating logs in absence of data. Alternative variables
may be required to accomplish specific goals, but will need to be
properly referenced when appropriate.
Code alignment to follow (Moving code left as a result of the above excess
logic change).