Add action for future implementation.
Idea is that Stock By Attributes provides multiple features, some of which
are additional options perhaps not used/needed by all and in some cases
desired to *not* be shown. Separating the display of the customid from
the operations of the core code would help those that wish to *not* display
that information and it should also help consideration(s) for other
development. E.g. if all of this can be done with/for this data, then couldn't
something else/more be done?

At any rate, consider ones self notified that there will be a change to
split out the display of the customid at least on the catalog side and may
even include an admin switch/change to adjust display for admin.