Remove template files that do not require modification.
Action to present the customid to the customer has predominantly
been moved to `includes/classes/observers/class.products_with_attributes_stock.php`
and therefore, a majority of previously touched template files no longer
need modification to support display of the customid.

This simplifies installation, though it does in some way bury the
content to be revised if necessary/desired.

Expectation is that the code to provide the customid on the catalog
side will in some way move to another observer so that one can
decide to not include it if desired. I expect that there will be
portions of the customid determination to remain in the core
SBA code.