Add array handling to eliminate need to evaluate function parameters
One way to fully evaluate the fed information is to use extra
parameters in function calling; however, this feature is becoming
more difficult to use in more recent PHP versions and requires a
lot of processing. Alternatively, where necessary, as in within
the code that is specific to this plugin, the information can be
combined in a way to pass the information and still allow results.

One benefit of using the method to look back on the function calls
is that for example, it could be possible to trigger the need to
display an "out of stock" message by reading the quantity of product
that were requested and could therefore force the software to
"detect" that the quantity remaining was negative even if that was
not at all true.