Revise method to align more with Admin side
Helps with display of customid information when reviewing order information.
For catalog side it does not merge the customid with other data. That is
to be done at the point of display or determined to display the information
(template related even if through an observer). Idea is that in design,
do not want to store the customid information directly with the product info
but instead in its own storage location to be combined as/where needed.

On the admin side, the data typically can be combined on the fly without
concern of affecting downstream data, generally as the items have already
been purchased.  That may not be the same with say Edit Orders or similar.