Add initial visualization of excess product.
At on point in time, SBA did not satisfactorily remove SBA variants
when attributes, product, etc. were deleted. As a result, product
may "appear" to be tracked by SBA; however, would be limited by
the lesser of product quantity of either the variant's quantity
or the total product quantity. This is/was furthered by the ability
of the store to "oversell" so it may report instances of being
out-of-stock when it isn't, or worse appear to be out-of-stock when
it would seem that the total product quantity indicated that it
was available.