Single commit to: Add additional "insert" with customid logic
Revised some of the handling logic for insertion operations when customid is present to assist in determining if duplication exists.

Add notification information for duplicate customid when duplication is not permitted.
Handling of a duplicate customid is not yet fully incorporated.
Some next steps:
Add handling of the extra/additional customids if when attempting to insert a customid that is expected to be unique that there are multiple matching customids.
External to the area touched, need a process to investigate the existence of duplicates when it has been determined that duplicates are not permitted.
  One idea here, is to return two types of information, that an update was performed and the need to update other records.