Expand to address for array as default
This section of code is used to alert and further describe when there
are elements in the database that cause product to incorrectly be seen
as stocked by attributes as in the product no longer have attributes or
the connection to them has been broken. Unfortunately, when there are/were
no such items to display, an empty array was being sent which while isset
was not what was desired for at either of the first two tests that are
changed.