Add unedited potential reporting tool
Thanks to fbroz for providing this reporting tool in the forum thread; https://www.zen-cart.com/showthread.php?47180-Stock-by-Attribute-v4-0-addon-for-v1-3-5-1-3-9&p=1333862#post1333862

This has been unedited from its original provision and has been untested by me. Brief review does indicate that there are things to do to expand on it, but at least I wanted to load it here to make it easier to find.

Some things that need to be done include adding a reference/link to load this file either in the SBA code and/or the admin screen in general. It may also be that some sort of hook is needed in an existing reporting tool in order to capture/use the functionality provided. More to follow at some point.