Address ZC1.5.8 function removal
Because of the renaming of functions in Zen Cart 1.5.8, the only identified
problem was this one where `zen_draw_products_pull_down` was renamed to
`zen_draw_pulldown_products`.

This solution supports backwards compatibility with predilection for the
more recent function as the older function is deprecated and expected to
generate a log on each use.