Adjust expected installation requirements.
As additional changes to the core code have been made to further support
installation without adding some specific files, the reporting of what
is needed/expected can be modified/reduced to apply to only earlier versions
of ZC. These changes help to better identify the files that are truly needed
for core operation, but identifying their absence does not prevent operation.
This is merely a check that is performed to provide information. It also
is not "highly complex" of a check, though it could be made more so.