Allow database insert without code failure if a duplicate was found.
For "upgrade" (currently a removal followed by an installation) ensuring that
the configuration keys that are to be added are flagged for deletion (or
prevent attempting to insert the key not deleted). This commit does two
things, it prevents failure of installation if the key already exists but
also removes the key(s) that should have already been part of the removal
process.

Fixes #105