Refactor display of some array assignments
Chose not to end array lists with a comma because this code is used in
older versions of PHP that may fail if the last item includes a
comma separator.

Also this wasn't done for all such possible cases, but was an observed
difference in files and it kind of stuck out to me.