Incorporate potential use of ZC 1.5.8 function(s)
Added a test that if function does not exist then use the old method of
testing if the constant is defined and then defining it if it isn't.

If the function does exist then use it to attempt to define the variable
if it is not yet defined.

Reasoning for capturing the test this way is that in the future, the code
could be truncated/modified to exclude the extra "older" code.