Restore ZC1.5.8 functionality accessing protected members
Changes made to the ZC 1.5.8 `shoppingCart` class reassigned some variables
from generically `public` to specifically `protected`.
SBA's `includes/extra_cart_actions`were developed from the perspective
of interacting with/feeding the `shoppingCart` class as the session
variable `$_SESSION['cart']`.

By making the below member variables `protected`, the `extra_cart_actions`
are no longer able to utilize operations existing in the `shoppingCart`
class and some of the benefits made available for communicating to the
customer.

This issue has been resolved by using class dependency injection to offer
handlers affecting the base values of the cart session. The use of
inheritence (`extends`) is not required to implement this concept, but I
wanted to leave behind some semblence of what was done here.

The `shoppingCart` class members that became protected in ZC 1.5.8 are:
These have get equivalent member functions to retrieve the data:
$content_type
$free_shipping_item
$free_shipping_weight
$free_shipping_price
$download_count
$total_before_discounts

These have get and set member functions to interact with the member:
$display_debug_messages
$flag_duplicate_msgs_set

This has each a get and add accessor member function:
$flag_duplicate_quantity_msgs_set