PHP 8.2/ZC 1.5.8 correction
created a debug log similar to:

````php
--> PHP Deprecated: str_replace(): Passing null to parameter #2 ($replace) of type array|string is deprecated in admin/products_with_attributes_stock.php on line 410.
--> PHP Deprecated: mysqli_real_escape_string(): Passing null to parameter #2 ($string) of type string is deprecated in catalog/includes/classes/db/mysql/query_factory.php on line 179.
````
The second item is related to the issue(s) with $skuTitle.
the `prepare_input` function is used elsewhere within the file; however,
have not yet gotten a report that there is an issue with the way it is
otherwise used/coded. Had planned to copy this change throughout the file
but didn't want to make changes where they were unnecessary.