Incorporate zen_define_default
Added code to use the ZC 1.5.8 function zen_define_default.
Unfortunately in this case, use of the function will assign a value as
determined by a ZC function that throws an error in this case.

This code originates from `admin/includes/modules/products_previous_next.php`

Brought up as a ZC issue, that appears will not be resolved/corrected at
least it will not look anything like it has/does.