Modify and refactor to focus on redirect `$parameters`
Add search parameters to link.

Further, in incorporation of the the search parameters to page redirects,
an opportunity was presented to refactor and focus on just the values that
need to be modified in association to the impending redirect. Both paths
required a redirect where the difference was only in the list of options in
`$parameters`.