ZC 1.5.8 Support continued use of `$db->getBindVarValue`
In Zen Cart 1.5.8, multiple methods and attributes of the `queryClass`
were removed from public access, with no handler implemented. By
extending to that class, methods within can access and use
the associated information.

This commit allows continued use of the `getBindVarValue` method as
specifically incorporated into this code and as described in the forum
for this plugin.