Add function to evaluate the `LIMIT` of a query.
In ZC 1.5.8, the newly formatted `split_page_results` class no longer
strips the limit portion of the query before adding its own. This addition
of a `LIMIT` to a query that has a `LIMIT` results in a query error.

This function evaluates `splitPageResults` class to determine suitability
to keep/remove the `LIMIT` from the query.