Push posted search to get.
Much of the search content is pulled from the browser link, therefore,
posted and get data should end up in a common location. Here that has been
made to be `$_GET['search']`.

This first attempts to collect `$_GET['search']`, then if `$_POST['search']`
is set, `$_GET['search']` is reassigned to `$_POST['search']`.